- Adding a validation set to tune hyperparameters. - Adding code for incorporating 1000G data as an LD reference panel

+++ data_preparation/generate_qc_filters.py
@@ -93,7 +93,7 @@ covar_df.to_csv(covar_file, sep="\t", header=False, index=False)
 
 def humanize_number(n):
     if n >= 1000:
-        return str(n / 1000) + 'k'
+        return str(int(n / 1000)) + 'k'
     else:
         return str(n)
